# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: gsm-system API v1
  version: 2.0.0
  description: |
    Standard REST API for the stateless GSM lab-cell toolkit (OpenBTS +
    Asterisk + smqueue). Legacy ad-hoc routes (/start, /stop, ...) remain
    frozen for compatibility; new integrations must use /api/v1.
servers:
  - url: http://<server>:8082
    description: SDR host (replace with yours)
# NOTE: the server runs OPEN by default (no token). Only when GSM_API_TOKEN
# is set must callers send `Authorization: Bearer <token>` on every call.
# JSON request bodies contain exactly one object. Cell start is limited to 1 MiB;
# the other JSON endpoints are limited to 64 KiB, including trailing data.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Send only when GSM_API_TOKEN is set on the server; otherwise omit.
...
paths:
...
  /api/v1/config:
    get:
      summary: Full OpenBTS config
      responses:
        '200': { description: '{config[]}' }
        '404': { description: OpenBTS database not found }
        '500': { description: config query failed }
    patch:
      summary: Update single config key (cell must be stopped)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: GSM.Identity.ShortName }
                value: { type: string, example: lab }
      responses:
        '200': { description: updated }
        '400': { description: malformed body }
        '404': { description: OpenBTS database not found }
        '409': { description: cell running }
        '413': { description: request body exceeds 64 KiB }
        '422': { description: invalid name/value }
        '500': { description: config update failed }
...